poll-backend-api/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
psycopg2-binary>=2.9.5
python-dotenv>=0.21.0
